requirements.txt
sklearn
tensorflow-gpu==1.12.2
redis
freegames